Add files via upload
Variant 1 of universe compression. It is based on the ideea that we don't have to "write" all the "particles" in a "zone" to have the whole picture. So.. we "compress" them to smaller mathematical/physics/programming notations. This first prototype, tries to simulate "filters","phenomena" over an unimaginable sized memory needs otherwise.